+++ x/fungible/keeper/grpc_query_foreign_coins.go
@@ -62,3 +62,21 @@ func (k Keeper) ForeignCoins(
 
 	return &types.QueryGetForeignCoinsResponse{ForeignCoins: val}, nil
 }
+
+// ForeignCoinsFromAsset returns the foreign coin for a given asset and chain id
+func (k Keeper) ForeignCoinsFromAsset(
+	c context.Context,
+	req *types.QueryGetForeignCoinsFromAssetRequest,
+) (*types.QueryGetForeignCoinsFromAssetResponse, error) {
+	if req == nil {
+		return nil, status.Error(codes.InvalidArgument, "invalid request")
+	}
+	ctx := sdk.UnwrapSDKContext(c)
+
+	fCoin, found := k.GetForeignCoinFromAsset(ctx, req.Asset, req.ChainId)
+	if !found {
+		return nil, status.Error(codes.NotFound, "not found")
+	}
+
+	return &types.QueryGetForeignCoinsFromAssetResponse{ForeignCoins: fCoin}, nil
+}