+++ zetaclient/logs/fields.go
@@ -3,19 +3,20 @@ package logs
 // A group of predefined field keys and module names for zetaclient logs
 const (
 	// field keys
-	FieldModule       = "module"
-	FieldMethod       = "method"
-	FieldChain        = "chain"
-	FieldChainNetwork = "chain_network"
-	FieldNonce        = "nonce"
-	FieldTracker      = "tracker_id"
-	FieldTx           = "tx"
-	FieldOutboundID   = "outbound_id"
-	FieldBlock        = "block"
-	FieldCctx         = "cctx"
-	FieldZetaTx       = "zeta_tx"
-	FieldBallot       = "ballot"
-	FieldCoinType     = "coin_type"
+	FieldModule           = "module"
+	FieldMethod           = "method"
+	FieldChain            = "chain"
+	FieldChainNetwork     = "chain_network"
+	FieldNonce            = "nonce"
+	FieldTracker          = "tracker_id"
+	FieldTx               = "tx"
+	FieldOutboundID       = "outbound_id"
+	FieldBlock            = "block"
+	FieldCctx             = "cctx"
+	FieldZetaTx           = "zeta_tx"
+	FieldBallot           = "ballot"
+	FieldCoinType         = "coin_type"
+	FieldConfirmationMode = "confirmation_mode"
 
 	// module names
 	ModNameInbound  = "inbound"